+++ src/OrbModding.Common/PluginIds.cs
@@ -16,11 +16,11 @@ public static class PluginIds
 
     public const string ModConfigGuid = "dev.vojow.orbofcreation.modconfig";
     public const string ModConfigName = "Orb Mod Config";
-    public const string ModConfigVersion = "0.5.2";
+    public const string ModConfigVersion = "0.5.3";
 
     public const string MentorGuid = "dev.vojow.orbofcreation.mentor";
     public const string MentorName = "Orb Mentor";
-    public const string MentorVersion = "0.1.1";
+    public const string MentorVersion = "0.1.2";
 
     public const string Version = "0.1.0";
 }